pyproject.toml
[project]
name = "user"
version = "0.0.1"
authors = [{ name = "Ramiro e Vinicius", email = "fulano@pucrs.com" }]
description = "A FastAPI project for user management"
readme = "README.md"
requires-python = ">=3.12"
keywords = ["user"]
dependencies = [
    "fastapi~=0.111",
    "loguru~=0.7",
    "pydantic[email]>=2,<3",
    "pydantic-settings",
    "sse-starlette>=2,<3",
    "uvicorn==0.30.3",
    "sqlalchemy>=1.4,<2.0",
    "beanie==1.26.0",
    "fastapi-cli==0.0.4",
    "httpx==0.27.0",
    "jinja2==3.1.4",
    "lazy-model==0.2.0",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "motor==3.5.1",
    "pymongo==4.8.0",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.1",
    "rich==13.7.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.37.2",
    "toml==0.10.2",
    "typer==0.12.3",
    "typing-extensions==4.12.2",
    "uvloop==0.19.0",
    "watchfiles==0.22.0",
    "websockets==12.0",
    "google-auth==2.3.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
addopts = [
    "--import-mode=importlib",
    "--cov=src",
    "--cov-report=html",
    "--disable-warnings",
]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
